move format fn to utils
