+++ src/analytics/index.ts
@@ -39,7 +39,6 @@ export function sendAnalyticsEvent(eventName: string, eventProperties?: Record<s
     console.log(`[analytics(${eventName})]: ${JSON.stringify(eventProperties)}`)
     return
   }
-
   track(eventName, eventProperties)
 }
