+++ src/components/Widget/transactions.ts
@@ -10,7 +10,7 @@ import { sendAnalyticsEvent } from 'analytics'
 import { EventName, SectionName } from 'analytics/constants'
 import { useTrace } from 'analytics/Trace'
 import { formatToDecimal, getTokenAddress } from 'analytics/utils'
-import { formatSwapSignedAnalyticsEventProperties } from 'components/swap/ConfirmSwapModal'
+import { formatSwapSignedAnalyticsEventProperties } from 'analytics/utils'
 import { WrapType } from 'hooks/useWrapCallback'
 import { useCallback, useMemo } from 'react'
 import { useTransactionAdder } from 'state/transactions/hooks'
